reorg test with utils
